pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "package-name"
# version = "0.1.0"
dynamic = ["version"]
description = "A sample Python package"
readme = "README.md"
authors = [
    { name = "Your Name", email = "your.email@example.com" },
]
maintainers = [
    { name = "Your Name", email = "your.email@example.com" },
]
license = { file = "LICENSE" }
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Education",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Recognition",
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS",
    "Operating System :: Microsoft :: Windows",
]
keywords = ["origin", "python3"]
requires-python = ">=3.8"
dependencies = [
    "loguru==0.7.3",
    "python-dotenv==1.0.0",
]

[tool.setuptools.dynamic]
version = { attr = "package_name.__version__" }

[project.urls]
"Homepage" = "https://pyorigin.com"
"Source" = "https://github.com/mokira3d48/PyOrigin/"
"Documentation" = "https://docs.pyorigin.com"
"Bug Reports" = "https://github.com/mokira3d48/PyOrigin/issues"
"Changelog" = "https://github.com/mokira3d48/PyOrigin/releases"

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
mycmd = "package_name.entrypoints.main:main"
